+++ src/main/java/babysnoozer/handlers/SnoozingBabyStateMachine.java
@@ -2,77 +2,74 @@ package babysnoozer.handlers;
 
 import babysnoozer.handlers.commands.CycleQueue;
 
-/**
- * Created by Alexander Bischof on 12.01.15.
- */
 public enum SnoozingBabyStateMachine {
 
-  SnoozingBabyStateMachine;
+    SnoozingBabyStateMachine;
 
-  public enum State {
-	Null, SetCycleCount, Snooze, ShutDown
-  }
+    public enum State {
+        Null, SetCycleCount, Snooze, ShutDown
+    }
 
-  private State state = State.Null;
+    private State state = State.Null;
 
-  private RopeDistance ropeDistance;
-  private int cycleCount = 1;
-  private CycleQueue cycles;
+    private RopeDistance ropeDistance;
+    private int cycleCount = 1;
+    private CycleQueue cycles;
 
-  public void setReleaseWaitTime(long releaseWaitTime) {
-    this.releaseWaitTime = releaseWaitTime;
-  }
+    public void setReleaseWaitTime(long releaseWaitTime) {
+        this.releaseWaitTime = releaseWaitTime;
+    }
 
-  public long getReleaseWaitTime() {
-    return releaseWaitTime;
-  }
+    public long getReleaseWaitTime() {
+        return releaseWaitTime;
+    }
 
-  private long releaseWaitTime = 120000l;
+    private long releaseWaitTime = 120000l;
 
-  public State getState() {
-	return state;
-  }
+    public State getState() {
+        return state;
+    }
 
-  public int getCycleCount() {
-	return cycleCount;
-  }
+    public int getCycleCount() {
+        return cycleCount;
+    }
 
-  public void setState(State state) {
-	this.state = state;
-  }
+    public void setState(State state) {
+        this.state = state;
+    }
 
-  public void setCycleCount(int cycleCount) {
-	this.cycleCount = cycleCount;
-  }
+    public void setCycleCount(int cycleCount) {
+        this.cycleCount = cycleCount;
+    }
 
-  public short getStartPos() {
-	return getRopeDistance().getStartPos();
-  }
+    public int getStartPos() {
+        return getRopeDistance().getStartPos();
+    }
 
-  public RopeDistance getRopeDistance() {
-	if (ropeDistance == null) {
-	  ropeDistance = new RopeDistance();
-	}
-	return ropeDistance;
-  }
+    public RopeDistance getRopeDistance() {
+        if (ropeDistance == null) {
+            ropeDistance = new RopeDistance();
+        }
+        return ropeDistance;
+    }
 
-  public CycleQueue getCycles() {
-	return cycles;
-  }
+    public CycleQueue getCycles() {
+        return cycles;
+    }
 
-  public void setCycles(CycleQueue cycles) {
-	this.cycles = cycles;
-  }
+    public void setCycles(CycleQueue cycles) {
+        this.cycles = cycles;
+    }
 
-  public void setStartPos(short startPos) {
-	getRopeDistance().setStartPos(startPos);
-  }
+    public void setStartPos(int startPos) {
+        getRopeDistance().setStartPos(startPos);
+    }
 
-  public short getEndPos() {
-	return getRopeDistance().getEndPos();
-  }
+    public int getEndPos() {
+        return getRopeDistance().getEndPos();
+    }
 
-  public void setEndPos(short endPos) {
-	getRopeDistance().setEndPos(endPos);
-  }
+    public void setEndPos(int endPos) {
+        getRopeDistance().setEndPos(endPos);
+    }
 }
